requirements.txt
python_pptx==0.6.18
numpy==1.13.3
